requirements.txt
aiogram
python-dotenv
sqlalchemy
psycopg2-binary
alembic